Temporary fix for gpuset.sh
Temporary resolution for an issue where gpuset.sh fails to shut down the display manager if vfio-pci is bound to any other device, which happens in some cases.
Currently the display manager will *always* be restarted if the service is restarted while /run/libvirt/request-passthrough exists. A more robust solution would be to check lspci not just for vfio-pci being used at all, but *specifically* on a VGA controller.